% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_funs.R
\name{convert_to_iupac}
\alias{convert_to_iupac}
\title{Convert tpo genotyping csv to iupac string}
\usage{
convert_to_iupac(tpocsv)
}
\arguments{
\item{tpocsv}{path to csv from tpo genotyping}
}
\value{
A chr of the IUPAC string
}
\description{
Convert tpo genotyping csv to iupac string
}
