% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_funs.R
\name{geno_comp}
\alias{geno_comp}
\title{Compare one genotype to many}
\usage{
geno_comp(goi, pop, filter = 0)
}
\arguments{
\item{goi}{a (named) genotype string in IUPAC format}
}
\value{
out A dataframe with the names and scores of all comparisons
}
\description{
Compare one genotype to many
}
